# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ArcherPilot AI API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all projects for the authenticated user
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, deployed, archived]
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Created project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a project by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateProject
      tags: [projects]
      summary: Update a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdate"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /projects/{id}/duplicate:
    post:
      operationId: duplicateProject
      tags: [projects]
      summary: Duplicate a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Duplicated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}/versions:
    get:
      operationId: listProjectVersions
      tags: [projects]
      summary: List version history for a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of versions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectVersion"
  /projects/{id}/export:
    post:
      operationId: exportProject
      tags: [projects]
      summary: Export a project in the specified format
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExportInput"
      responses:
        "200":
          description: Export result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExportResult"

  # ── AI ─────────────────────────────────────────────────────────────────────
...
  /deployments:
    get:
      operationId: listDeployments
      tags: [deployments]
      summary: List all deployments for the authenticated user
      responses:
        "200":
          description: List of deployments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Deployment"
    post:
      operationId: startDeployment
      tags: [deployments]
      summary: Start a deployment (simulated if no Archer connection)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeploymentInput"
      responses:
        "201":
          description: Started deployment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deployment"
  /deployments/{id}:
    get:
      operationId: getDeployment
      tags: [deployments]
      summary: Get a deployment by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deployment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deployment"

  # ── Dashboard ───────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]

    # ── Project schemas ──────────────────────────────────────────────────────
    Project:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        name:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [draft, deployed, archived]
        prompt:
          type: ["string", "null"]
        content:
          oneOf:
            - $ref: "#/components/schemas/ArcherImplementation"
            - type: "null"
        versionCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, userId, name, status, versionCount, createdAt, updatedAt]
    ProjectInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        prompt:
          type: string
        content:
          $ref: "#/components/schemas/ArcherImplementation"
      required: [name]
    ProjectUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        status:
          type: string
          enum: [draft, deployed, archived]
        content:
          $ref: "#/components/schemas/ArcherImplementation"
    ProjectVersion:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        versionNumber:
          type: integer
        content:
          oneOf:
            - $ref: "#/components/schemas/ArcherImplementation"
            - type: "null"
        createdAt:
          type: string
          format: date-time
      required: [id, projectId, versionNumber, createdAt]
    ExportInput:
      type: object
      properties:
        format:
          type: string
          enum: [word, pdf, markdown, json, archer-package]
      required: [format]
    ExportResult:
      type: object
      properties:
        format:
          type: string
        filename:
          type: string
        content:
          type: string
        mimeType:
          type: string
      required: [format, filename, content, mimeType]

    # ── Archer Implementation (AI-generated structure) ───────────────────────
    ArcherImplementation:
      type: object
      properties:
        businessOverview:
          type: string
        useCases:
          type: array
          items:
            type: string
        applicationStructure:
          type: object
          additionalProperties: true
        modules:
          type: array
          items:
            type: object
            additionalProperties: true
        fields:
          type: array
          items:
            type: object
            additionalProperties: true
        valueLists:
          type: array
          items:
            type: object
            additionalProperties: true
        crossReferences:
          type: array
          items:
            type: object
            additionalProperties: true
        questionnaires:
          type: array
          items:
            type: object
            additionalProperties: true
        workflow:
          type: object
          additionalProperties: true
        recordPermissions:
          type: array
          items:
            type: object
            additionalProperties: true
        notifications:
          type: array
          items:
            type: object
            additionalProperties: true
        reports:
          type: array
          items:
            type: object
            additionalProperties: true
        dashboards:
          type: array
          items:
            type: object
            additionalProperties: true
        sampleRecords:
          type: array
          items:
            type: object
            additionalProperties: true
        testCases:
          type: array
          items:
            type: object
            additionalProperties: true

    # ── AI schemas ───────────────────────────────────────────────────────────
...
    Deployment:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        userId:
          type: string
        status:
          type: string
          enum: [pending, running, completed, failed, simulated]
        steps:
          type: array
          items:
            $ref: "#/components/schemas/DeploymentStep"
        error:
          type: ["string", "null"]
        connectionId:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, projectId, userId, status, steps, createdAt, updatedAt]
    DeploymentStep:
      type: object
      properties:
        name:
          type: string
        status:
          type: string
          enum: [pending, running, completed, failed]
        message:
          type: ["string", "null"]
      required: [name, status]
    DeploymentInput:
      type: object
      properties:
        projectId:
          type: integer
        connectionId:
          type: ["integer", "null"]
        simulate:
          type: boolean
      required: [projectId]

    # ── Dashboard schemas ────────────────────────────────────────────────────
...