# lib/api-zod/src/generated/types/deploymentInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * ArcherPilot AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DeploymentInput {
  projectId: number;
  /** @nullable */
  connectionId?: number | null;
  simulate?: boolean;
}